requirements.txt
line-bot-sdk==3.5.0
flask
apscheduler
python-3.11.9
